fix(security): widen integer index operands in convolve/moment/psnr preserving float SIMD bit-exactness

+++ core/src/feature/iqa/convolve.c
@@ -131,7 +131,7 @@ static void iqa_convolve_horizontal_pass(const float *img, int w, const struct i
             const int kx = x + uc;
             const ptrdiff_t img_offset = (ptrdiff_t)ky * w + kx;
             for (int u = -uc; u <= uc - kw_even; ++u, ++k_offset) {
-                sum += (double)img[img_offset + u] * k->kernel_h[k_offset];
+                sum += img[img_offset + u] * k->kernel_h[k_offset];
             }
             img_cache[img_offset] = (float)(sum * scale);
         }
@@ -152,9 +152,9 @@ static void iqa_convolve_vertical_pass(const float *img_cache, int w, const stru
             const int kx = x + uc;
             const ptrdiff_t img_offset = (ptrdiff_t)ky * w + kx;
             for (int v = -vc; v <= vc - kh_even; ++v, ++k_offset) {
-                sum += (double)img_cache[img_offset + (ptrdiff_t)v * w] * k->kernel_v[k_offset];
+                sum += img_cache[img_offset + (ptrdiff_t)v * w] * k->kernel_v[k_offset];
             }
-            dst[y * dst_w + x] = (float)(sum * scale);
+            dst[(ptrdiff_t)y * dst_w + x] = (float)(sum * scale);
         }
     }
 }
@@ -292,10 +292,9 @@ float iqa_filter_pixel(const float *img, int w, int h, int x, int y, const struc
         const ptrdiff_t img_offset = (ptrdiff_t)(y + v) * w + x;
         for (int u = -uc; u <= uc - kw_even; ++u, ++k_offset) {
             if (!edge) {
-                sum += (double)img[img_offset + u] * k->kernel[k_offset];
+                sum += img[img_offset + u] * k->kernel[k_offset];
             } else {
-                sum +=
-                    (double)k->bnd_opt(img, w, h, x + u, y + v, k->bnd_const) * k->kernel[k_offset];
+                sum += k->bnd_opt(img, w, h, x + u, y + v, k->bnd_const) * k->kernel[k_offset];
             }
         }
     }
