+++ core/src/feature/moment.c
@@ -37,7 +37,7 @@ int compute_1st_moment(const float *pic, int w, int h, int stride, double *score
 
     for (int i = 0; i < h; ++i) {
         for (int j = 0; j < w; ++j) {
-            pic_ = pic[i * stride_ + j];
+            pic_ = pic[(ptrdiff_t)i * stride_ + j];
             cum += pic_;
         }
     }
@@ -57,8 +57,8 @@ int compute_2nd_moment(const float *pic, int w, int h, int stride, double *score
 
     for (int i = 0; i < h; ++i) {
         for (int j = 0; j < w; ++j) {
-            pic_ = pic[i * stride_ + j];
-            cum += (double)pic_ * pic_;
+            pic_ = pic[(ptrdiff_t)i * stride_ + j];
+            cum += pic_ * pic_;
         }
     }
     cum /= ((double)w * h);
