+++ core/src/feature/psnr.c
@@ -42,10 +42,10 @@ int compute_psnr(const float *ref, const float *dis, int w, int h, int ref_strid
 
     for (int i = 0; i < h; ++i) {
         for (int j = 0; j < w; ++j) {
-            float ref_ = ref[i * ref_stride_ + j];
-            float dis_ = dis[i * dis_stride_ + j];
+            float ref_ = ref[(ptrdiff_t)i * ref_stride_ + j];
+            float dis_ = dis[(ptrdiff_t)i * dis_stride_ + j];
             float diff = ref_ - dis_;
-            noise_ += (double)diff * diff;
+            noise_ += diff * diff;
         }
     }
     noise_ /= ((double)w * h);
